added more error handling and broke up models and controllers
