+++ app.js
@@ -1,10 +1,13 @@
+const { response } = require("express");
 const express = require("express");
 const app = express();
 const {
 	getCategories,
+} = require("./controllers/categoriesControllers");
+const {
 	getReviews,
 	getReviewsById,
-} = require("./controllers/controllers");
+}=require('./controllers/ReviewsControllers')
 
 app.get("/api/categories", getCategories);
 
@@ -12,10 +15,21 @@ app.get("/api/reviews", getReviews);
 
 app.get("/api/reviews/:review_id", getReviewsById);
 
+app.use((error, request, response, next) => {
+	if(error.code='22P02'){
+	console.log(error);
+	response.status(400).send({ msg: "bad request" });
+	}else{
+		next(error)
+	}
+});
+
+app.use((error, request, response, next) => {
+	console.log(error);
+	response.status(500).send({ msg: "server error" });
+});
+
 app.use("*", (request, response, next) => {
 	response.status(404).send({ msg: "not found" });
-});
-app.use((error, request, response, next) => {
-	response.status(500).send({ msg: "server error" });
 });
 module.exports = app;
