+++ models/models.js
@@ -5,9 +5,3 @@ exports.fetchData=(input)=>{
 		return result.rows
 	})
 }
-
-exports.fetchReviewsById=(id)=>{
-	return db.query('SELECT * FROM reviews WHERE review_id=$1',[id]).then((result)=>{
-		return result.rows
-	})
-}
